fakenect: don't exit(1) if FAKENECT_PATH not set
freenect_init() is designed to be able to report errors gracefully by
returning -1 and applications may implement graceful fallbacks if they
can't initialize libfreenect. A fallback is not possible if fakenect exits
the program when FAKENECT_PATH is unset. The application may not be in
a good position to set FAKENECT_PATH automatically to avoid the exit()
given that it doesn't typically know when fakenect is in use.